Ported BitSetStepper (new name) to standard library.
Algorithm has changed; new one should be faster for dense and/or small
bitsets (arguably more important), and anyway is easier to verify
for correctness.

Also is better about not holding on to the original BitSet when it's no
longer needed.

Note that because the element type is manually specialized, we have both
a direct (simple) method for `stepper` as well as the elaborated Shape
version that manually picks out whether the correct shape is requested
so it can use its efficient implementation instead of the superclass
implementation.