Added Stepper to Vector.
Stepper gets direct access, outside of `VectorPointer`, to the internal
arrays that back `Vector`.  This is a lot easier now that Stepper lives
inside the main library; `Vector` can hand off the requisite arrays
without exposing its internals to the world.

Note also that the arrays are not fully formed when the vector is `dirty`.
Rather than trying to deal with dirty arrays it just uses cleaned versions
from an `Iterator`.  This is slightly inefficient, but as Vector isn't
the most high-performance data structure anyway when used with single-element
addition (which is where things get dirty), it shouldn't matter.

Also removed some unnecessary imports.